Machine Learning/requirements.txt
fastapi==0.75.0
numpy==1.22.2
Pillow==9.0.1
tensorflow-cpu==2.8.0
gunicorn==20.1.0
uvicorn==0.17.6
python-multipart==0.0.5